requirements.txt
absl-py @ file:///C:/ci/absl-py_1615411407614/work
aiohttp @ file:///C:/ci/aiohttp_1614361133907/work
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
Babel==2.9.0
backcall==0.2.0
black==20.8b1
bleach==3.3.0
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet==4.0.0
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
colorama==0.4.4
coverage @ file:///C:/ci/coverage_1614614910274/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
cycler==0.10.0
Cython @ file:///C:/ci/cython_1614015037049/work
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
entrypoints==0.3
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1616008050444/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1614894617465/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1614884419385/work
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1615900693889/work
ipykernel==5.5.0
ipyparallel==6.3.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
ke==0.0.2
Keras==2.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
Markdown @ file:///C:/ci/markdown_1614364005059/work
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.4.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607367865052/work
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nose==1.3.7
notebook==6.3.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1615224334542/work
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.8.1
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pypiwin32==223
pyreadline==2.1
pyrsistent==0.17.3
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.18.1
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.1
six @ file:///C:/ci/six_1605205426665/work
snowballstemmer==2.1.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
tifffile==2021.3.17
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
widgetsnbextension==3.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
